examples/requirements.txt
streamlit >= 0.81
streamlit-yellowbrick